model_service/requirements.txt
grpcio
grpcio-tools
protobuf
torch
transformers
sentencepiece
